requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
cachetools==5.3.2
python-socketio==5.10.0
eventlet==0.34.2
transformers>=4.36.0
torch>=2.0.0
pillow>=10.0.0
